Remove CrossIcon and Update CloseIcon
